removed template module
