--- template/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-from template.models import Template
-
-# Register your models here.
-admin.site.register(Template)
